create report today and report by date create transaction flow

+++ dto/product.go
@@ -3,7 +3,7 @@ package dto
 type CreateProduct struct {
 	Name       string  `json:"name"`
 	Price      float64 `json:"price"`
-	Stock      int     `json:"stock"`
+	Stock      float64 `json:"stock"`
 	CategoryId string  `json:"category_id"`
 }
 
@@ -11,5 +11,5 @@ type UpdateProduct struct {
 	ID    string  `json:"id"`
 	Name  string  `json:"name"`
 	Price float64 `json:"price"`
-	Stock int     `json:"stock"`
+	Stock float64 `json:"stock"`
 }
